mv cipher_mapping bits all to cipher_util
