+++ lib/ansible/parsing/vault/__init__.py
@@ -19,7 +19,6 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-from collections import defaultdict
 import os
 import shlex
 import shutil
@@ -31,8 +30,7 @@ from subprocess import call
 from ansible.errors import AnsibleError, AnsibleVaultError
 from ansible.parsing.vault import envelope
 
-from ansible.plugins import cipher_loader
-from ansible.parsing.vault.cipher_util import CIPHER_ENCRYPT_WHITELIST, get_decrypt_cipher, get_encrypt_cipher
+from ansible.parsing.vault import cipher_util
 
 from ansible.module_utils._text import to_bytes
 
@@ -74,40 +72,6 @@ def is_encrypted_file(file_obj, start_pos=0, count=-1):
         file_obj.seek(current_position)
 
 
-# Which is better is a judgement call, and here is how we rank our crypto children
-cipher_impl_preference = {'cryptography': 100,
-                          'PyCrypto': 50,
-                          'other': 10,
-                          'no_impl': 0,
-                          "Larry's Original Crypto Shack and Shoe Repair": -1}
-
-
-def get_impl_score(cph_class):
-    # if plugin doesnt provide implementation attr, then it gets 0 points. But
-    # we can still rank something lower...
-    impl = getattr(cph_class, 'implementation', 'no_impl')
-    # and default to 0 for unknown impl as well
-    score = cipher_impl_preference.get(impl, 0)
-    return score
-
-
-def get_cipher_mapping():
-    cipher_class_gen = cipher_loader.all(class_only=True)
-
-    # map cipher_name -> list of classes that provide it
-    mapping = defaultdict(list)
-    for cc in cipher_class_gen:
-        mapping[cc.name].append(cc)
-
-    cipher_mapping = {}
-
-    # set our prefered cipher to be the class with the highest score (last in the list when sorted via get_score())
-    for c_name, c_classes in mapping.items():
-        cipher_mapping[c_name] = sorted(c_classes, key=get_impl_score).pop()
-
-    return cipher_mapping
-
-
 class VaultLib:
 
     # The prereqs can be different for each cipher impl
@@ -121,7 +85,7 @@ class VaultLib:
         self.b_version = b'1.1'
 
         # TODO: could do this later? lazily?
-        self.cipher_mapping = get_cipher_mapping()
+        self.cipher_mapping = cipher_util.get_cipher_mapping()
 
     @staticmethod
     def is_encrypted(data):
@@ -161,7 +125,7 @@ class VaultLib:
             raise AnsibleVaultError("input is already encrypted")
 
         # use the default cipher
-        cipher_class = get_encrypt_cipher(self.default_cipher_name, self.cipher_mapping)
+        cipher_class = cipher_util.get_encrypt_cipher(self.default_cipher_name, self.cipher_mapping)
         cipher = cipher_class()
 
         # encrypt data
@@ -197,7 +161,7 @@ class VaultLib:
 
     def decrypt_ciphertext(self, b_ciphertext, cipher_name, b_version):
 
-        cipher_class = get_decrypt_cipher(cipher_name, self.cipher_mapping)
+        cipher_class = cipher_util.get_decrypt_cipher(cipher_name, self.cipher_mapping)
         cipher = cipher_class()
 
         # try to unencrypt vaulttext
@@ -389,7 +353,7 @@ class VaultEditor:
         except AnsibleError as e:
             raise AnsibleVaultError("%s for %s" % (to_bytes(e), to_bytes(filename)))
 
-        if cipher_name not in CIPHER_ENCRYPT_WHITELIST:
+        if cipher_name not in cipher_util.CIPHER_ENCRYPT_WHITELIST:
             # we want to get rid of files encrypted with the AES cipher
             self._edit_file_helper(filename, existing_data=plaintext, force_save=True)
         else:
